moved network config to a separate file
separated cc tokenv2 commands from the token sample

readme updates

addded forgotten params for cc functions

moving configs towards user
