+++ .gitignore
@@ -4,3 +4,4 @@ node_modules
 npm-debug.log
 .idea
 *.iml
+config/index.js
