+++ config/index.sample.js
@@ -1,0 +1,6 @@
+module.exports = {
+  /** TOKENS V2 config */
+  // tokens v2 global pubkey:
+  tokensv2GlobalPk:
+    '032fd27f72591b02f13a7f9701246eb0296b2be7cfdad32c520e594844ec3d4801',
+};
