+++ src/networks.js
@@ -42,8 +42,7 @@ function getDefaultBip32Testnet () {
   }
 }
 
-module.exports = {
-
+module.exports =  {
   // https://github.com/bitcoin/bitcoin/blob/master/src/validation.cpp
   // https://github.com/bitcoin/bitcoin/blob/master/src/chainparams.cpp
   bitcoin: {
@@ -313,25 +312,5 @@ module.exports = {
     coin: coins.ZEC,
     komodoAssetNet: true,
     magic: 0x794cdf9b //   0x9bdf4c79
-  },
-  
-  tkltest: {
-    messagePrefix: '\x18TKLTEST asset chain:\n',
-    bech32: 'R',
-    bip32: getDefaultBip32Mainnet(),
-    pubKeyHash: 0x3c,
-    scriptHash: 0x55,
-    wif: 0xbc,
-    consensusBranchId: {
-      1: 0x00,
-      2: 0x00,
-      3: 0x5ba81b19,
-      4: 0x76b809bb // (old Sapling branch id - used in kmd)
-      // 4: 0x2bb40e60
-      // 4: 0xf5b9230b
-    },
-    coin: coins.ZEC,
-    komodoAssetNet: true,
-    magic:  0xf6475548 
-  },
+  }
 }
